requirements.txt
numpy==1.11.0
scipy==0.17.1
cycler==0.10.0
dask==0.9.0
decorator==4.0.9
mahotas==1.4.1
matplotlib==1.5.1
networkx==1.11
pandas==0.18.1
Pillow==3.2.0
pymorph==0.96
pyparsing==2.1.4
python-dateutil==2.5.3
pytz==2016.4
scikit-image==0.12.3
scikit-learn==0.17.1
six==1.10.0
toolz==0.8.0
wheel==0.29.0
